Move timeout classes to timeouts.py
